+++ src/Profile/Shopware/Exception/MigrationShopwareProfileException.php
@@ -14,33 +14,7 @@ use Symfony\Component\HttpFoundation\Response;
 #[Package('services-settings')]
 class MigrationShopwareProfileException extends HttpException
 {
-    public const ASSOCIATION_ENTITY_REQUIRED_MISSING = 'SWAG_MIGRATION__SHOPWARE_ASSOCIATION_REQUIRED_MISSING';
-
-    public const DATABASE_CONNECTION_ERROR = 'SWAG_MIGRATION__DATABASE_CONNECTION_ERROR';
-
     public const PLUGIN_NOT_INSTALLED = 'SWAG_MIGRATION__PLUGIN_NOT_INSTALLED';
-
-    public static function associationEntityRequiredMissing(string $entity, string $missingEntity): self
-    {
-        return new AssociationEntityRequiredMissingException(
-            Response::HTTP_NOT_FOUND,
-            self::ASSOCIATION_ENTITY_REQUIRED_MISSING,
-            'Mapping of "{{ missingEntity }}" is missing, but it is a required association for "{{ entity }}". Import "{{ missingEntity }}" first.',
-            [
-                'missingEntity' => $missingEntity,
-                'entity' => $entity,
-            ]
-        );
-    }
-
-    public static function databaseConnectionError(): self
-    {
-        return new self(
-            Response::HTTP_INTERNAL_SERVER_ERROR,
-            self::DATABASE_CONNECTION_ERROR,
-            'Database connection could not be established.'
-        );
-    }
 
     public static function pluginNotInstalled(): self
     {