+++ src/Profile/Shopware/Gateway/Local/ShopwareLocalGateway.php
@@ -17,12 +17,12 @@ use Shopware\Core\System\Currency\CurrencyEntity;
 use Shopware\Core\System\Language\LanguageCollection;
 use Shopware\Core\System\Language\LanguageEntity;
 use Shopware\Core\System\Locale\LocaleEntity;
+use SwagMigrationAssistant\Exception\MigrationException;
 use SwagMigrationAssistant\Migration\EnvironmentInformation;
 use SwagMigrationAssistant\Migration\Gateway\Reader\EnvironmentReaderInterface;
 use SwagMigrationAssistant\Migration\Gateway\Reader\ReaderRegistry;
 use SwagMigrationAssistant\Migration\MigrationContextInterface;
 use SwagMigrationAssistant\Migration\RequestStatusStruct;
-use SwagMigrationAssistant\Profile\Shopware\Exception\MigrationShopwareProfileException;
 use SwagMigrationAssistant\Profile\Shopware\Gateway\Connection\ConnectionFactoryInterface;
 use SwagMigrationAssistant\Profile\Shopware\Gateway\ShopwareGatewayInterface;
 use SwagMigrationAssistant\Profile\Shopware\Gateway\TableReaderInterface;
@@ -75,7 +75,7 @@ class ShopwareLocalGateway implements ShopwareGatewayInterface
         $profile = $migrationContext->getProfile();
 
         if ($connection === null) {
-            $error = MigrationShopwareProfileException::databaseConnectionError();
+            $error = MigrationException::databaseConnectionError();
 
             return new EnvironmentInformation(
                 $profile->getSourceSystemName(),
@@ -90,7 +90,7 @@ class ShopwareLocalGateway implements ShopwareGatewayInterface
         try {
             $connection->connect();
         } catch (\Throwable $e) {
-            $error = MigrationShopwareProfileException::databaseConnectionError();
+            $error = MigrationException::databaseConnectionError();
 
             return new EnvironmentInformation(
                 $profile->getSourceSystemName(),